news_summarize/requirements.txt
streamlit==1.35.0
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.20
langchain-google-genai==0.0.6
python-dotenv==1.0.1
